I'm timid to speak but willing to converse about the good tidings. I'm a musical fanatic and spend most of my years writing and arranging my own lyrics. I belongto the faith of my ancestors and present my angelic form as a nazarite to the world of israelites, the enlightened ones. I'm modest and wise enough to be more than the average individual as the components that build on intelligence makes technology exiting whiles life is less exciting for others. I'm here in this world to be the diction of others left in the dark of this powerful instument.
Programming the future for music means including the software and data science into a culture rich in the art of music. I'm a local musician with the aspiring thoughts of the greatest names to have come so far. Hampered by growth, developing speed and the mechanics of musical software, the soul feel and power of real music is being missed single mindedly and the content of the new age writers sounds more like garbage as the delineate from their morals and rights. Consciousness is my great will to exercise upon humanity the power of love and with given time and opportunities in the technical world will open the consciousness of the spiritual ones and we can enjoy a world of peace all together.